+++ src/Service/Ses/src/ValueObject/Template.php
@@ -5,7 +5,8 @@ namespace AsyncAws\Ses\ValueObject;
 /**
  * An object that defines the email template to use for an email message, and the values to use for any message
  * variables in that template. An *email template* is a type of message template that contains content that you want to
- * define, save, and reuse in email messages that you send.
+ * reuse in email messages that you send. You can specifiy the email template by providing the name or ARN of an *email
+ * template* previously saved in your Amazon SES account or by providing the full template content.
  */
 final class Template
 {
@@ -23,6 +24,16 @@ final class Template
      * @var string|null
      */
     private $templateArn;
+
+    /**
+     * The content of the template.
+     *
+     * > Amazon SES supports only simple substitions when you send email using the `SendEmail` or `SendBulkEmail` operations
+     * > and you provide the full template content in the request.
+     *
+     * @var EmailTemplateContent|null
+     */
+    private $templateContent;
 
     /**
      * An object that defines the values to use for message variables in the template. This object is a set of key-value
@@ -44,6 +55,7 @@ final class Template
      * @param array{
      *   TemplateName?: null|string,
      *   TemplateArn?: null|string,
+     *   TemplateContent?: null|EmailTemplateContent|array,
      *   TemplateData?: null|string,
      *   Headers?: null|array<MessageHeader|array>,
      * } $input
@@ -52,6 +64,7 @@ final class Template
     {
         $this->templateName = $input['TemplateName'] ?? null;
         $this->templateArn = $input['TemplateArn'] ?? null;
+        $this->templateContent = isset($input['TemplateContent']) ? EmailTemplateContent::create($input['TemplateContent']) : null;
         $this->templateData = $input['TemplateData'] ?? null;
         $this->headers = isset($input['Headers']) ? array_map([MessageHeader::class, 'create'], $input['Headers']) : null;
     }
@@ -60,6 +73,7 @@ final class Template
      * @param array{
      *   TemplateName?: null|string,
      *   TemplateArn?: null|string,
+     *   TemplateContent?: null|EmailTemplateContent|array,
      *   TemplateData?: null|string,
      *   Headers?: null|array<MessageHeader|array>,
      * }|Template $input
@@ -80,6 +94,11 @@ final class Template
     public function getTemplateArn(): ?string
     {
         return $this->templateArn;
+    }
+
+    public function getTemplateContent(): ?EmailTemplateContent
+    {
+        return $this->templateContent;
     }
 
     public function getTemplateData(): ?string
@@ -103,6 +122,9 @@ final class Template
         }
         if (null !== $v = $this->templateArn) {
             $payload['TemplateArn'] = $v;
+        }
+        if (null !== $v = $this->templateContent) {
+            $payload['TemplateContent'] = $v->requestBody();
         }
         if (null !== $v = $this->templateData) {
             $payload['TemplateData'] = $v;